backend/requirements.txt
Fastapi==0.101.1
uvicorn==0.23.2
aiohttp==3.9.1
PyJWT==2.8.0
python-multipart==0.0.6
openapi_spec_validator==0.7.1
APScheduler==3.10.4
python-dotenv==1.0.0
gunicorn==21.2.0
asyncio==3.4.3
python_dateutil==2.8.2

# crypto
bcrypt==4.0.1

# db
asyncpg==0.28.0
aioredis==2.0.1

# text_splitter
tiktoken==0.5.2
